// src/types/content.ts
import { FirebaseTimestampType } from '../firebase'
import { ImageType } from './image'
import { PeriodType } from './period'

export type ContentType = {
  admin_id: string
  content_id: string
  created_at?: FirebaseTimestampType
  description: string
  genre: string
  image: ImageType[]
  period: PeriodType
  skills: string[]
  title: string
  updated_at?: FirebaseTimestampType
  url?: string
}

export type HistoryType = {
  admin_id: string
  created_at?: FirebaseTimestampType
  event: string
  history_id: string
  month: number
  year: number
  // role:
  updated_at?: FirebaseTimestampType
}

export type ProfileType = {
  admin_id: string
  created_at: FirebaseTimestampType
  description: string
  email: string
  image: ImageType
  name: string
  updated_at: FirebaseTimestampType
}